+++ scout/ext/backend/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function, absolute_import, unicode_literals
+from __future__ import (print_function, absolute_import, unicode_literals)
+
 from .base import BaseAdapter
 from .fixture import FixtureAdapter
 from .vcf import VcfAdapter