+++ scout/models/case/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import unicode_literals
+
+STATUS = (
+  "active",
+  "inactive",
+  "solved",
+  "archived"
+)