+++ scout/models/case/case.py
@@ -1,86 +1,25 @@
-# -*- coding: utf-8 -*-
-"""
-
-"main concept of MongoDB is embed whenever possible"
-Ref: http://stackoverflow.com/questions/4655610#comment5129510_4656431
-"""
-from __future__ import absolute_import, unicode_literals
+from __future__ import (unicode_literals, absolute_import)
 from datetime import datetime
 import itertools
 
-from query_phenomizer import query
-from mongoengine import (DateTimeField, Document, EmbeddedDocument,
-                         EmbeddedDocumentField, IntField, ListField,
-                         ReferenceField, FloatField, StringField,
-                         BooleanField)
+from mongoengine import (Document, StringField, ListField, ReferenceField,
+                         EmbeddedDocumentField, DateTimeField, BooleanField,
+                         BinaryField, FloatField)
+import query_phenomizer
 
-from .event import Event
-
-STATUS_ORDER = ['inactive', 'active', 'archived', 'solved']
-
-
-class Individual(EmbeddedDocument):
-  """Represents an individual (sample) in a case (family)."""
-  display_name = StringField()
-  sex = StringField()
-  phenotype = IntField()
-  father = StringField()
-  mother = StringField()
-  individual_id = StringField()
-  capture_kits = ListField(StringField())
-  bam_file = StringField()
-
-  @property
-  def sex_human(self):
-    """Transform sex string into human readable form."""
-    # pythonic switch statement
-    return {'1': 'male', '2': 'female'}.get(self.sex, 'unknown')
-
-  @property
-  def phenotype_human(self):
-    """Transform phenotype integer into human readable form."""
-    # pythonic switch statement
-    terms = {-9: 'missing', 0: 'missing', 1: 'unaffected', 2: 'affected'}
-    return terms.get(self.phenotype, 'undefined')
-
-  def __unicode__(self):
-    return self.display_name
-
-  def __repr__(self):
-    return "Individual(individual_id={0}, display_name={1})".format(
-      self.individual_id, self.display_name
-    )
-
-
-class PhenotypeTerm(EmbeddedDocument):
-  hpo_id = StringField()
-  feature = StringField()
-
-  def __repr__(self):
-    return "PhenotypeTerm(hpo_id={0}, feature={1})".format(
-      self.hpo_id, self.feature
-    )
-
-
-class GeneList(EmbeddedDocument):
-  list_id = StringField(required=True)
-  version = FloatField(required=True)
-  date = StringField(required=True)
-  display_name = StringField()
-
-  def __repr__(self):
-    return "GeneList(list_id={0}, version={1}, date={2}, display_name={3})".format(
-      self.list_id, self.version, self.date, self.display_name
-    )
+from . import STATUS
+from .individual import Individual
+from .gene_list import GeneList
+from scout.models import (User, Variant, PhenotypeTerm)
 
 
 class Case(Document):
   """Represents a case (family) of individuals (samples)."""
-  # This is the md5 string id for the family:
+  # This is a string with the id for the family:
   case_id = StringField(primary_key=True, required=True)
   # This is the string that will be shown in scout:
   display_name = StringField(required=True)
-  # This is the owner of the case
+  # This is the owner of the case. E.g. 'cust003'
   owner = StringField(required=True)
   # These are the names of all the collaborators that are allowed to view the
   # case, including the owner
@@ -91,16 +30,18 @@ class Case(Document):
   updated_at = DateTimeField(default=datetime.now)
   suspects = ListField(ReferenceField('Variant'))
   causative = ReferenceField('Variant')
-  synopsis = StringField(default='')
-  status = StringField(default='inactive', choices=STATUS_ORDER)
-  is_research = BooleanField()
-  events = ListField(EmbeddedDocumentField(Event))
-  comments = ListField(EmbeddedDocumentField(Event))
 
-  # This decides which gene lists that should be shown when the case is opened
+  # The synopsis is a text blob
+  synopsis = StringField(default='')
+  status = StringField(default='inactive', choices=STATUS)
+  is_research = BooleanField(default=False)
+
+  # default_gene_lists specifies which gene lists that should be shown when
+  # the case is opened
   default_gene_lists = ListField(StringField())
   clinical_gene_lists = ListField(EmbeddedDocumentField(GeneList))
   research_gene_lists = ListField(EmbeddedDocumentField(GeneList))
+  dynamic_gene_list = ListField(StringField())
 
   genome_build = StringField()
   genome_version = FloatField()
@@ -114,7 +55,8 @@ class Case(Document):
   madeline_info = StringField()
   vcf_file = StringField()
 
-  coverage_report_path = StringField()
+  # The coverage report will be read as a binary blob
+  coverage_report = BinaryField()
 
   @property
   def is_solved(self):
@@ -145,7 +87,7 @@ class Case(Document):
     # skip querying Phenomizer unless at least one HPO terms exists
     if hpo_terms:
       try:
-        return query(hpo_terms)
+        return query_phenomizer.query(hpo_terms)
       except SystemExit:
         return {}
     else:
@@ -169,8 +111,4 @@ class Case(Document):
 
   def __repr__(self):
     return "Case(case_id={0}, display_name={1}, owner={2})".format(
-      self.case_id, self.display_name, self.owner
-    )
-
-  # def __unicode__(self):
-  #   return self.display_name
+      self.case_id, self.display_name, self.owner)